+++ cmd/local/conf/parse_config.go
@@ -23,13 +23,15 @@ import (
 	"gopkg.in/yaml.v3"
 )
 
-func ParseConfig(configDir string, overrides map[string]string) (map[string]interface{}, error) {
-
+func ParseConfig(ddcYamlLoc string, overrides map[string]string) (map[string]interface{}, error) {
 	data := make(map[string]interface{})
-	expectedLoc := filepath.Join(configDir, "ddc.yaml")
-	confFile, err := os.ReadFile(filepath.Clean(expectedLoc))
+	absPath, err := filepath.Abs(ddcYamlLoc)
 	if err != nil {
-		return data, fmt.Errorf("conf %v not found, and cannot read directory %v due to error %w", expectedLoc, configDir, err)
+		return data, fmt.Errorf("can't absolute path for %v due to error %w", ddcYamlLoc, err)
+	}
+	confFile, err := os.ReadFile(filepath.Clean(absPath))
+	if err != nil {
+		return data, fmt.Errorf("conf %v not readable due to error %w", ddcYamlLoc, err)
 	}
 
 	err = yaml.Unmarshal(confFile, &data)
@@ -38,7 +40,7 @@ func ParseConfig(configDir string, overrides map[string]string) (map[string]inte
 		return data, fmt.Errorf("unable to parse yaml: %w", err)
 	}
 
-	simplelog.Debugf("conf %v parsed successfully", expectedLoc)
+	simplelog.Infof("conf %v parsed successfully", absPath)
 
 	for k, v := range overrides {
 		//this really only applies for running over ssh so why am I doing it here? because we end up doing some crazy stuff as a result!