+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ### Added
 * ddc.log file location on command start and on command end
+* ddc falls back to logging to the temp directory if the default location is not present
+* --ddc-yaml flag now for local-collect to be able to read a ddc yaml from anywhere
 
 ## [0.7.4]
 ### Fixed